requirements.txt
Django-4.1 asgiref-3.7.2 django-braces-1.15.0